[refactor: authorizationSetting]
