+++ client/src/features/Staff/index.js
@@ -1,0 +1,12 @@
+import { Typography } from '@mui/material';
+import React from 'react';
+
+function Staff(props) {
+  return (
+    <Typography variant="h4" gutterBottom>
+      Nhân Viên
+    </Typography>
+  );
+}
+
+export default Staff;
