+++ client/src/utils/getComparator.js
@@ -1,0 +1,7 @@
+import descendingComparator from 'utils/descendingComparator';
+
+export default function getComparator(order, orderBy) {
+  return order === 'desc'
+    ? (a, b) => descendingComparator(a, b, orderBy)
+    : (a, b) => -descendingComparator(a, b, orderBy);
+}
